+++ src/forze/application/contracts/document/__init__.py
@@ -4,14 +4,17 @@ from .deps import (
     DocumentQueryDepKey,
     DocumentQueryDepPort,
 )
+from .ensure_id import assert_unique_ensure_ids, require_create_id_for_ensure
 from .ports import DocumentCommandPort, DocumentQueryPort
 from .specs import DocumentSpec, DocumentWriteTypes
 
 # ----------------------- #
 
 __all__ = [
+    "assert_unique_ensure_ids",
     "DocumentQueryPort",
     "DocumentCommandPort",
+    "require_create_id_for_ensure",
     "DocumentSpec",
     "DocumentWriteTypes",
     "DocumentQueryDepKey",